requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.2
Deprecated==1.3.1
Flask==3.1.3
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
Mako==1.3.11
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.1
PyJWT==2.12.1
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
wrapt==2.1.2
